LMB-519: move /util/types.ts to root file types.ts
